Hackathon2_scripts/Hackathon2_scripts/requirements.txt
flask
flask-cors
numpy
pillow
opencv-python-headless
torch
torchvision
ultralytics
requests